requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==2.0.1
transformers==4.30.2
pytube==15.0.0
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==9.5.0 